refactor(backend): clean up files with llm and webapp main packages
