+++ backend/src/llm/llms.py
@@ -1,5 +1,5 @@
 from langchain_google_vertexai import VertexAI
-from modules.config import GCP_PROJECT
+from llm.config import GCP_PROJECT
 
 chat_bison = VertexAI(model_name="text-bison", project=GCP_PROJECT)
 code_bison_6k = VertexAI(model_name="code-bison", project=GCP_PROJECT)
