+++ backend/src/llm/static/schema-description.txt
